# docs/swagger.yml
openapi: 3.0.0
info:
  description: |
    SIH Backend Server Documentation
  version: "1.0.0"
  title: SIH-Backend
  contact:
    email: raghhavdturki@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - description: Development Server
    url: http://localhost:7071/api/
  - description: Production Server
    url: https://sih-2022-server.azurewebsites.net/api/
  - description: Chat Server
    url: https://sih-email.herokuapp.com/api/chat
paths:
...
  /messages:
    get:
      tags:
        - chat
      security:
        - bearerAuth: []
      summary: Fetch all the messages in a room
      operationId: messages
      parameters: 
        - in: query
          name: roomId
          schema:
            type: string
          required: true
      responses:
        '400':
          description: Invalid input
        '500':
          description: Internal Server Error
        '200':
          description: Messages fetched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetMessagesResponse'
...
  /conversation:
    get:
      tags:
        - chat
      security:
        - bearerAuth: []
      summary: Fetch a conversation from a documentId
      operationId: conversation
      parameters: 
        - in: query
          name: documentId
          schema:
            type: string
          required: true
      responses:
        '400':
          description: Invalid input
        '500':
          description: Internal Server Error
        '200':
          description: Conversation fetched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
...
components:
  schemas:
...
    Conversation:
      type: object
      properties:
        conversationName:
          type: string
        documentId:
          type: string
        participants:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              info:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
    Message:
      type: object
      properties:
        senderId:
          type: string
        subject:
          type: string
        content:
          type: string
        conversationId:
          type: string
        senderName: 
          type: string
        createdAt: 
          type: string
          format: date-time
        updatedAt: 
          type: string
          format: date-time
...
    GetMessagesResponse:
      type: array
      items:
        $ref: "#/components/schemas/Message"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Iron Sealed Token
...